# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeReconOps API - Cross-border financial reconciliation platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /upload/bank:
    post:
      operationId: uploadBankStatement
      tags: [upload]
      summary: Upload bank statement CSV
      description: Upload and parse a USD bank statement CSV file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Invalid CSV format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /upload/ledger:
    post:
      operationId: uploadLedger
      tags: [upload]
      summary: Upload internal ledger CSV
      description: Upload and parse a CNY internal ledger CSV file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Upload successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Invalid CSV format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /results:
    get:
      operationId: getResults
      tags: [results]
      summary: Get reconciliation results
      description: Returns paginated, filtered reconciliation results
      parameters:
        - name: batchId
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: status
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: riskLevel
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          required: false
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Reconciliation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultsPage"
  /results/{resultId}:
    get:
      operationId: getResultById
      tags: [results]
      summary: Get single reconciliation result
      parameters:
        - name: resultId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Single result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReconciliationResult"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]
      required:
        - error
    UploadResult:
      type: object
      properties:
        batchId:
          type: string
        rowCount:
          type: integer
        fileName:
          type: string
        uploadType:
          type: string
        errors:
          type: array
          items:
            type: string
        warnings:
          type: array
          items:
            type: string
      required:
        - batchId
        - rowCount
        - fileName
        - uploadType
        - errors
        - warnings
...
    ReconciliationResult:
      type: object
      properties:
        id:
          type: string
        batchId:
          type: string
        bankDate:
          type: ["string", "null"]
        bankDescription:
          type: ["string", "null"]
        bankAmount:
          type: ["string", "null"]
        bankCurrency:
          type: ["string", "null"]
        bankType:
          type: ["string", "null"]
        bankReference:
          type: ["string", "null"]
        ledgerDate:
          type: ["string", "null"]
        ledgerVendorClient:
          type: ["string", "null"]
        ledgerAmount:
          type: ["string", "null"]
        ledgerCurrency:
          type: ["string", "null"]
        ledgerInvoiceId:
          type: ["string", "null"]
        ledgerCategory:
          type: ["string", "null"]
        exchangeRate:
          type: ["string", "null"]
        convertedAmount:
          type: ["string", "null"]
        differenceAmount:
          type: ["string", "null"]
        differencePct:
          type: ["string", "null"]
        status:
          type: string
          enum:
            - MATCHED
            - FX_VARIANCE
            - AMOUNT_MISMATCH
            - UNMATCHED_BANK
            - UNMATCHED_LEDGER
            - POSSIBLE_DUPLICATE
            - MANUAL_REVIEW
        riskLevel:
          type: string
          enum:
            - LOW
            - MEDIUM
            - HIGH
            - CRITICAL
        reason:
          type: string
        fuzzyScore:
          type: ["number", "null"]
      required:
        - id
        - batchId
        - status
        - riskLevel
        - reason
...
    ResultsPage:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/ReconciliationResult"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        totalPages:
          type: integer
      required:
        - results
        - total
        - page
        - pageSize
        - totalPages
...