# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeReconOps API - Cross-border financial reconciliation platform
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  batchId: string;
  rowCount: number;
  fileName: string;
  uploadType: string;
  errors: string[];
  warnings: string[];
}